// src/app/enums/logs.enum.ts
export enum ActionTypeEnum {
  Hold = "Hold",
  Active = "Active",
  Resticit = "Resticit",
  Extend = "Extend",
  ChangeName = "ChangeName",
  ChangePhone = "ChangePhone",
  ChangeAdress = "ChangeAdress",
  Changedriver = "Changedriver",
  ChangeBranch = "ChangeBranch",
  AddNotes = "AddNotes",
  AddDays = "AddDays",
  AddMeals = "AddMeals",
  changeStartDate = "changeStartDate",
  Detact = "Detact",
  BagCount = "BagCount",
  Renew = "Renew",
  CustomMeals = "CustomMeals",
  Migrations = "Migrations",
  ChangeState = "ChangeState",
  ReplaceMeals = "ReplaceMeals",
}